refactored web controller
