+++ src/main/java/ru/legionofone/klassikaplusserver/web/controller/RevisionController.java
@@ -1,38 +1,43 @@
 package ru.legionofone.klassikaplusserver.web.controller;
 
-
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.ui.Model;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ru.legionofone.klassikaplusserver.service.CatalogService;
 
-import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
+import java.util.List;
 
-@RestController
+@RestController("revision")
 public class RevisionController {
 
-    final CatalogService service;
+    private static final Logger logger = LoggerFactory.getLogger(RevisionController.class.getName());
+
+    private CatalogService catalog;
 
     @Autowired
     public RevisionController(CatalogService service) {
-        this.service = service;
+        this.catalog = service;
     }
 
-    @GetMapping("revision")
-    public Integer getRevision(@RequestParam(required = false) String deviceId) {
-        return service.getRevision();
+    @GetMapping
+    public Integer getCurrentRevision() {
+        logger.info("Providing revision number");
+        return catalog.getRevision();
     }
 
-    @GetMapping("Greeting")
-    public String greeting(@RequestParam(name = "name", required = false, defaultValue = "Anonymous") String name, Model model){
-        model.addAttribute("name", model);
-        return "greeting";
-    }
 
-    @GetMapping("admin")
-    public String admin(HttpServletRequest request){
-        return "admin";
+    @PostMapping
+    public ResponseEntity updateRevision() {
+        logger.info("Request updating revision");
+        catalog.updateRevision();
+
+        List<Object> list = new ArrayList<>();
+
+        return ResponseEntity.ok().build();
     }
 }
